moved EnvoirmentHelper to Infrastructure.Core
