+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/System/HealthStatus.cs
@@ -20,6 +20,7 @@ using System.Threading.Tasks;
 using Azure.Messaging.ServiceBus.Administration;
 using GreenEnergyHub.Charges.FunctionHost.Common;
 using GreenEnergyHub.Charges.FunctionHost.Configuration;
+using GreenEnergyHub.Charges.Infrastructure.Core.Registration.ServiceCollection;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Data.SqlClient;
